feat: cross-platform build with OpenSSL-based JWT
- Implement JWT (HS256) directly using OpenSSL (~150 lines)
- Remove libjwt dependency (v3 API incompatibility)
- Fix 30+ cross-platform compiler warnings (sign conversions, format specifiers)
- Support aggressive warnings (-Werror) on both GCC and Clang
- Add proper library linking for macOS (pkg-config directories)
- Fix atomic type casting in circuit_breaker.c
- Use PRIu64 format specifiers for uint64_t in clickhouse.c
- Add const correctness to clickhouse_config_t fields
- Suppress false positives in logging.c and redis_publisher.c
- Platform-specific compiler flags (GCC vs Clang)

Tested on:
- AppleClang 17.0.0 (macOS arm64)
- JWT implementation validated with 5 test cases

All 4 services compile cleanly with -Werror:
- ethhook-admin-api: 323KB
- ethhook-ingestor: 303KB
- ethhook-processor: 306KB
- ethhook-delivery: 71KB

+++ ethhook-c/src/common/arena.c
@@ -45,7 +45,7 @@ void *eth_arena_alloc(eth_arena_t *arena, size_t size) {
     }
     
     // Align to 8 bytes
-    size = (size + 7) & ~7;
+    size = (size + 7) & ~((size_t)7);
     
     pthread_mutex_lock(&arena->lock);
